Further cleanup (#638)

+++ src/arch/helperrvv.h
@@ -1171,7 +1171,7 @@ static INLINE void vscatter2_v_p_i_i_vd(double *ptr, int offset, int step, vdoub
   // probably only iterate 2 or 4 times.
   //
   ptr += offset * 2;
-  for (int i = 0; i < VECTLENDP; i += 2) {
+  for (int i = 0; i < (int)VECTLENDP; i += 2) {
     // PROTIP: Avoid modifying `v` within the loop, and just extract the useful
     // part directly in each iteration, because we can.  This avoids a
     // loop-carried dependency.
@@ -1185,7 +1185,7 @@ static INLINE void vscatter2_v_p_i_i_vd(double *ptr, int offset, int step, vdoub
 static INLINE void vscatter2_v_p_i_i_vf(float *ptr, int offset, int step, vfloat v) {
   // as above re: looping
   ptr += offset * 2;
-  for (int i = 0; i < VECTLENSP; i += 2) {
+  for (int i = 0; i < (int)VECTLENSP; i += 2) {
     vfloat vv = __riscv_vslidedown(v, i, 2);
     __riscv_vse32(ptr, vv, 2);
     ptr += step * 2;
